--- dokumentlager/domene/src/main/java/no/nav/foreldrepenger/melding/felles/ContainerLoginCDI.java
@@ -1,15 +1,0 @@
-package no.nav.foreldrepenger.melding.felles;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-import javax.inject.Qualifier;
-
-
-@Qualifier
-@Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.TYPE, ElementType.METHOD, ElementType.FIELD, ElementType.PARAMETER})
-public @interface ContainerLoginCDI {
-}
