+++ dokumentlager/domene/src/main/java/no/nav/foreldrepenger/melding/historikk/HistorikkRepositoryImpl.java
@@ -7,6 +7,7 @@ import javax.inject.Inject;
 import javax.persistence.EntityManager;
 import javax.persistence.TypedQuery;
 
+import no.nav.vedtak.felles.jpa.HibernateVerktøy;
 import no.nav.vedtak.felles.jpa.VLPersistenceUnit;
 
 @ApplicationScoped
@@ -33,6 +34,13 @@ public class HistorikkRepositoryImpl implements HistorikkRepository {
         TypedQuery<DokumentHistorikkinnslag> query = entityManager.createQuery("from DokumentHistorikkinnslag where behandlingId=:behandlingId", DokumentHistorikkinnslag.class);
         query.setParameter("behandlingId", behandlingId);
         return query.getResultList();
+    }
+
+    @Override
+    public DokumentHistorikkinnslag hentInnslagMedId(long historikkinnslagId) {
+        TypedQuery<DokumentHistorikkinnslag> query = entityManager.createQuery("from DokumentHistorikkinnslag where id=:historikkinnslagId", DokumentHistorikkinnslag.class);
+        query.setParameter("historikkinnslagId", historikkinnslagId);
+        return HibernateVerktøy.hentEksaktResultat(query);
     }
 
     private void lagreOgFlush(Object objektTilLagring) {
