+++ domenetjenester/brevbestiller/pom.xml
@@ -20,6 +20,10 @@
             <artifactId>domene</artifactId>
         </dependency>
         <dependency>
+            <groupId>no.nav.foreldrepenger.felles</groupId>
+            <artifactId>felles-prosesstask</artifactId>
+        </dependency>
+        <dependency>
             <groupId>no.nav.foreldrepenger.melding</groupId>
             <artifactId>fpsak</artifactId>
         </dependency>
@@ -55,6 +59,10 @@
             <artifactId>migreringer</artifactId>
             <scope>test</scope>
             <type>test-jar</type>
+        </dependency>
+        <dependency>
+            <groupId>no.nav.foreldrepenger.melding</groupId>
+            <artifactId>dokumenthistorikk</artifactId>
         </dependency>
     </dependencies>
 </project>
