+++ domenetjenester/brevbestiller/src/main/java/no/nav/foreldrepenger/melding/brevbestiller/task/ProduserBrevTaskProperties.java
@@ -1,0 +1,12 @@
+package no.nav.foreldrepenger.melding.brevbestiller.task;
+
+public class ProduserBrevTaskProperties {
+
+    public static final String TASKTYPE = "formidling.bestillBrev";
+
+    public static final String HENDELSE_ID = "hendelseId";
+
+    private ProduserBrevTaskProperties() {
+        // Skal ikke konstrueres
+    }
+}
