+++ domenetjenester/fpsak/src/main/java/no/nav/foreldrepenger/fpsak/BehandlingRestKlient.java
@@ -10,9 +10,9 @@ import no.nav.foreldrepenger.fpsak.dto.personopplysning.PersonopplysningDto;
 import no.nav.foreldrepenger.fpsak.dto.personopplysning.VergeDto;
 
 public interface BehandlingRestKlient {
-    Optional<BehandlingDto> hentBehandling(BehandlingIdDto behandlingIdDto, boolean systembruker);
+    Optional<BehandlingDto> hentBehandling(BehandlingIdDto behandlingIdDto);
 
-    Optional<PersonopplysningDto> hentPersonopplysninger(BehandlingIdDto behandlingIdDto, List<BehandlingResourceLinkDto> resourceLinkDtos, boolean systembruker);
+    Optional<PersonopplysningDto> hentPersonopplysninger(BehandlingIdDto behandlingIdDto, List<BehandlingResourceLinkDto> resourceLinkDtos);
 
-    Optional<VergeDto> hentVerge(BehandlingIdDto behandlingIdDto, List<BehandlingResourceLinkDto> resourceLinkDtos, boolean systembruker);
+    Optional<VergeDto> hentVerge(BehandlingIdDto behandlingIdDto, List<BehandlingResourceLinkDto> resourceLinkDtos);
 }
