+++ kafkatjenester/dokumenthendelse/src/main/java/no/nav/foreldrepenger/melding/kafkatjenester/dokumenthendelse/KafkaReader.java
