+++ kafkatjenester/dokumenthendelse/src/main/resources/META-INF/beans.xml
