+++ kafkatjenester/dokumenthistorikk/src/main/java/no/nav/foreldrepenger/melding/kafkatjenester/historikk/task/PubliserHistorikkTaskProperties.java
@@ -1,0 +1,10 @@
+package no.nav.foreldrepenger.melding.kafkatjenester.historikk.task;
+
+public class PubliserHistorikkTaskProperties {
+    public static final String TASKTYPE = "formidling.publiserHistorikk";
+    public static final String HISTORIKK_ID = "historikkId";
+
+    private PubliserHistorikkTaskProperties() {
+    }
+
+}
