+++ migreringer/src/main/resources/db/migration/defaultDS/1.0/V1.1_07__Tabell_dokumenthendelse_DDL.sql
@@ -7,23 +7,23 @@ CREATE TABLE DOKUMENT_HENDELSE
   ID                 bigint,
   BEHANDLING_ID      bigint                                 NOT NULL,
   BEHANDLING_TYPE    VARCHAR(100),
-  YTELSE_TYPE        VARCHAR(100)                           NOT NULL,
-  DOKUMENT_MAL_NAVN  VARCHAR(7),
-  GJELDER_VEDTAK     BOOLEAN      DEFAULT FALSE,
-  FRITEKST           TEXT,
-  TITTEL             VARCHAR(200),
-  HISTORIKK_AKTOER   VARCHAR(100),
-  OPPRETTET_AV       VARCHAR(20)  DEFAULT 'VL',
-  OPPRETTET_TID      TIMESTAMP(3) DEFAULT localtimestamp,
-  ENDRET_AV          VARCHAR(20),
-  ENDRET_TID         TIMESTAMP(3),
-  KL_BEHANDLING_TYPE VARCHAR(100) DEFAULT 'BEHANDLING_TYPE' NOT NULL,
-  KL_YTELSE_TYPE     VARCHAR(100) DEFAULT 'FAGSAK_YTELSE'   NOT NULL,
-  KL_HISTORIKK_AKTOR VARCHAR(100) DEFAULT 'FAGSAK_YTELSE'   NOT NULL,
+  YTELSE_TYPE         VARCHAR(100)                           NOT NULL,
+  DOKUMENT_MAL_NAVN   VARCHAR(7),
+  GJELDER_VEDTAK      BOOLEAN      DEFAULT FALSE,
+  FRITEKST            TEXT,
+  TITTEL              VARCHAR(200),
+  HISTORIKK_AKTOER    VARCHAR(100),
+  OPPRETTET_AV        VARCHAR(20)  DEFAULT 'VL',
+  OPPRETTET_TID       TIMESTAMP(3) DEFAULT localtimestamp,
+  ENDRET_AV           VARCHAR(20),
+  ENDRET_TID          TIMESTAMP(3),
+  KL_BEHANDLING_TYPE  VARCHAR(100) DEFAULT 'BEHANDLING_TYPE'  NOT NULL,
+  KL_YTELSE_TYPE      VARCHAR(100) DEFAULT 'FAGSAK_YTELSE'    NOT NULL,
+  KL_HISTORIKK_AKTOER VARCHAR(100) DEFAULT 'HISTORIKK_AKTOER' NOT NULL,
   CONSTRAINT FK_DOKUMENT_HENDELSE_1 FOREIGN KEY (DOKUMENT_MAL_NAVN) REFERENCES DOKUMENT_MAL_TYPE (KODE),
   CONSTRAINT FK_DOKUMENT_HENDELSE_2 FOREIGN KEY (BEHANDLING_TYPE, KL_BEHANDLING_TYPE) REFERENCES KODELISTE (KODE, KODEVERK),
   CONSTRAINT FK_DOKUMENT_HENDELSE_3 FOREIGN KEY (YTELSE_TYPE, KL_YTELSE_TYPE) REFERENCES KODELISTE (KODE, KODEVERK),
-  CONSTRAINT FK_DOKUMENT_HENDELSE_4 FOREIGN KEY (HISTORIKK_AKTOER, KL_HISTORIKK_AKTOR) REFERENCES KODELISTE (KODE, KODEVERK)
+  CONSTRAINT FK_DOKUMENT_HENDELSE_4 FOREIGN KEY (HISTORIKK_AKTOER, KL_HISTORIKK_AKTOER) REFERENCES KODELISTE (KODE, KODEVERK)
 );
 
 CREATE INDEX IDX_DOKUMENT_HENDELSE_1 on DOKUMENT_HENDELSE (BEHANDLING_ID);
@@ -44,5 +44,5 @@ COMMENT ON COLUMN DOKUMENT_HENDELSE.TITTEL IS 'Tittel for fritekstbrev';
 COMMENT ON COLUMN DOKUMENT_HENDELSE.HISTORIKK_AKTOER IS 'FK:HISTORIKK_AKTOER Fremmednøkkel til';
 COMMENT ON COLUMN DOKUMENT_HENDELSE.KL_YTELSE_TYPE IS 'Referanse til KODEVERK-kolonnen i KODELISTE-tabellen';
 COMMENT ON COLUMN DOKUMENT_HENDELSE.KL_BEHANDLING_TYPE IS 'Referanse til KODEVERK-kolonnen i KODELISTE-tabellen';
-COMMENT ON COLUMN DOKUMENT_HENDELSE.KL_HISTORIKK_AKTOR IS 'Referanse til KODEVERK-kolonnen i KODELISTE-tabellen';
+COMMENT ON COLUMN DOKUMENT_HENDELSE.KL_HISTORIKK_AKTOER IS 'Referanse til KODEVERK-kolonnen i KODELISTE-tabellen';
 COMMENT ON TABLE DOKUMENT_HENDELSE IS 'Hendelser om dokumenter bestilt eller forhåndvist fra foreldrepengeløsningen';
