+++ migreringer/src/main/resources/db/migration/defaultDS/1.0/V1.1_14__Prosesstask_DDL.sql
@@ -1,0 +1,143 @@
+CREATE TABLE PROSESS_TASK
+(
+  ID                        bigint,
+  TASK_TYPE                 VARCHAR(50),
+  PRIORITET                 integer      DEFAULT 0,
+  STATUS                    VARCHAR(20)  DEFAULT 'KLAR',
+  TASK_PARAMETERE           VARCHAR(4000),
+  TASK_PAYLOAD              text,
+  TASK_GRUPPE               VARCHAR(250),
+  TASK_SEKVENS              VARCHAR(100) DEFAULT '1',
+  NESTE_KJOERING_ETTER      TIMESTAMP(0) DEFAULT current_timestamp,
+  FEILEDE_FORSOEK           integer      DEFAULT 0,
+  SISTE_KJOERING_TS         TIMESTAMP(6),
+  SISTE_KJOERING_FEIL_KODE  VARCHAR(50),
+  SISTE_KJOERING_FEIL_TEKST text,
+  SISTE_KJOERING_SERVER     VARCHAR(50),
+  VERSJON                   bigint       DEFAULT 0,
+  OPPRETTET_AV              VARCHAR(20)  DEFAULT 'VL',
+  OPPRETTET_TID             TIMESTAMP(6) DEFAULT localtimestamp,
+  BLOKKERT_AV               bigint
+); /* TODO: Partisjonering  ENABLE__ROW MOVEMENT ; */
+
+COMMENT ON COLUMN PROSESS_TASK.ID IS 'Primary Key';
+COMMENT ON COLUMN PROSESS_TASK.TASK_TYPE IS 'navn på task. Brukes til å matche riktig implementasjon';
+COMMENT ON COLUMN PROSESS_TASK.PRIORITET IS 'prioritet på task.  Høyere tall har høyere prioritet';
+COMMENT ON COLUMN PROSESS_TASK.STATUS IS 'status på task: KLAR, NYTT_FORSOEK, FEILET, VENTER_SVAR, FERDIG';
+COMMENT ON COLUMN PROSESS_TASK.TASK_PARAMETERE IS 'parametere angitt for en task';
+COMMENT ON COLUMN PROSESS_TASK.TASK_PAYLOAD IS 'inputdata for en task';
+COMMENT ON COLUMN PROSESS_TASK.TASK_GRUPPE IS 'angir en unik id som grupperer flere ';
+COMMENT ON COLUMN PROSESS_TASK.TASK_SEKVENS IS 'angir rekkefølge på task innenfor en gruppe ';
+COMMENT ON COLUMN PROSESS_TASK.NESTE_KJOERING_ETTER IS 'tasken skal ikke kjøeres før tidspunkt er passert';
+COMMENT ON COLUMN PROSESS_TASK.FEILEDE_FORSOEK IS 'antall feilede forsøk';
+COMMENT ON COLUMN PROSESS_TASK.SISTE_KJOERING_TS IS 'siste gang tasken ble forsøkt kjørt';
+COMMENT ON COLUMN PROSESS_TASK.SISTE_KJOERING_FEIL_KODE IS 'siste feilkode tasken fikk';
+COMMENT ON COLUMN PROSESS_TASK.SISTE_KJOERING_FEIL_TEKST IS 'siste feil tasken fikk';
+COMMENT ON COLUMN PROSESS_TASK.SISTE_KJOERING_SERVER IS 'navn på node som sist kjørte en task (server@pid)';
+COMMENT ON COLUMN PROSESS_TASK.VERSJON IS 'angir versjon for optimistisk låsing';
+COMMENT ON COLUMN PROSESS_TASK.BLOKKERT_AV IS 'Id til ProsessTask som blokkerer kjøring av denne (når status=VETO)';
+COMMENT ON TABLE PROSESS_TASK IS 'Inneholder tasks som skal kjøres i bakgrunnen';
+
+CREATE TABLE PROSESS_TASK_FEILHAND
+(
+  KODE            VARCHAR(20),
+  NAVN            VARCHAR(50),
+  BESKRIVELSE     VARCHAR(2000),
+  OPPRETTET_AV    VARCHAR(20)  DEFAULT 'VL',
+  OPPRETTET_TID   TIMESTAMP(3) DEFAULT localtimestamp,
+  ENDRET_AV       VARCHAR(20),
+  ENDRET_TID      TIMESTAMP(3),
+  INPUT_VARIABEL1 bigint,
+  INPUT_VARIABEL2 bigint
+);
+
+COMMENT ON COLUMN PROSESS_TASK_FEILHAND.KODE IS 'Kodeverk Primary Key';
+COMMENT ON COLUMN PROSESS_TASK_FEILHAND.NAVN IS 'Lesbart navn på type feilhåndtering brukt i prosesstask';
+COMMENT ON COLUMN PROSESS_TASK_FEILHAND.BESKRIVELSE IS 'Utdypende beskrivelse av koden';
+COMMENT ON COLUMN PROSESS_TASK_FEILHAND.INPUT_VARIABEL1 IS 'Variabel 1: Dynamisk konfigurasjon for en feilhåndteringsstrategi.  Verdi og betydning er bestemt av feilhåndteringsalgoritmen';
+COMMENT ON COLUMN PROSESS_TASK_FEILHAND.INPUT_VARIABEL2 IS 'Variabel 2: Dynamisk konfigurasjon for en feilhåndteringsstrategi.  Verdi og betydning er bestemt av feilhåndteringsalgoritmen';
+COMMENT ON TABLE PROSESS_TASK_FEILHAND IS 'Kodetabell for feilhåndterings-metoder. For eksempel antall ganger å prøve på nytt og til hvilke tidspunkt';
+
+CREATE TABLE PROSESS_TASK_TYPE
+(
+  KODE                     VARCHAR(50),
+  NAVN                     VARCHAR(50),
+  FEIL_MAKS_FORSOEK        integer      DEFAULT 1,
+  FEIL_SEK_MELLOM_FORSOEK  integer      DEFAULT 30,
+  CRON_EXPRESSION          VARCHAR(200) NULL,
+  FEILHANDTERING_ALGORITME VARCHAR(20)  DEFAULT 'DEFAULT',
+  BESKRIVELSE              VARCHAR(2000),
+  OPPRETTET_AV             VARCHAR(20)  DEFAULT 'VL',
+  OPPRETTET_TID            TIMESTAMP(3) DEFAULT localtimestamp,
+  ENDRET_AV                VARCHAR(20),
+  ENDRET_TID               TIMESTAMP(3)
+);
+
+COMMENT ON COLUMN PROSESS_TASK_TYPE.KODE IS 'Kodeverk Primary Key';
+COMMENT ON COLUMN PROSESS_TASK_TYPE.NAVN IS 'Lesbart navn på prosesstasktype';
+COMMENT ON COLUMN PROSESS_TASK_TYPE.FEIL_MAKS_FORSOEK IS 'Maksimalt anntall forsøk på rekjøring om noe går galt';
+COMMENT ON COLUMN PROSESS_TASK_TYPE.CRON_EXPRESSION IS 'Cron-expression for når oppgaven skal kjøres på nytt';
+COMMENT ON COLUMN PROSESS_TASK_TYPE.FEIL_SEK_MELLOM_FORSOEK IS 'Ventetid i sekunder mellom hvert forsøk på rekjøring om noe har gått galt';
+COMMENT ON COLUMN PROSESS_TASK_TYPE.FEILHANDTERING_ALGORITME IS 'FK:PROSESS_TASK_FEILHAND Fremmednøkkel til tabell som viser detaljer om hvordan en feilsituasjon skal håndteres';
+COMMENT ON COLUMN PROSESS_TASK_TYPE.BESKRIVELSE IS 'Utdypende beskrivelse av koden';
+COMMENT ON TABLE PROSESS_TASK_TYPE IS 'Kodetabell for typer prosesser med beskrivelse og informasjon om hvilken feilhåndteringen som skal benyttes';
+
+CREATE SEQUENCE SEQ_PROSESS_TASK MINVALUE 1000000 MAXVALUE 999999999999999999 INCREMENT BY 50 START WITH 1032200 NO CYCLE;
+
+CREATE SEQUENCE SEQ_PROSESS_TASK_GRUPPE MINVALUE 10000000 MAXVALUE 999999999999999999 INCREMENT BY 1000000 START WITH 21384000000 NO CYCLE;
+
+
+ALTER TABLE PROSESS_TASK
+  ALTER COLUMN OPPRETTET_TID SET NOT NULL;
+ALTER TABLE PROSESS_TASK
+  ALTER COLUMN OPPRETTET_AV SET NOT NULL;
+ALTER TABLE PROSESS_TASK
+  ADD CONSTRAINT PK_PROSESS_TASK PRIMARY KEY (ID);
+ALTER TABLE PROSESS_TASK
+  ADD CONSTRAINT CHK_PROSESS_TASK_STATUS CHECK (STATUS in
+                                                ('KLAR', 'FEILET', 'VENTER_SVAR', 'SUSPENDERT', 'VETO', 'FERDIG'));
+ALTER TABLE PROSESS_TASK
+  ALTER COLUMN VERSJON SET NOT NULL;
+ALTER TABLE PROSESS_TASK
+  ALTER COLUMN TASK_SEKVENS SET NOT NULL;
+ALTER TABLE PROSESS_TASK
+  ALTER COLUMN STATUS SET NOT NULL;
+ALTER TABLE PROSESS_TASK
+  ALTER COLUMN PRIORITET SET NOT NULL;
+ALTER TABLE PROSESS_TASK
+  ALTER COLUMN TASK_TYPE SET NOT NULL;
+ALTER TABLE PROSESS_TASK
+  ALTER COLUMN ID SET NOT NULL;
+
+ALTER TABLE PROSESS_TASK_FEILHAND
+  ADD CONSTRAINT PK_PROSESS_TASK_FEILHAND PRIMARY KEY (KODE);
+ALTER TABLE PROSESS_TASK_FEILHAND
+  ALTER COLUMN OPPRETTET_TID SET NOT NULL;
+ALTER TABLE PROSESS_TASK_FEILHAND
+  ALTER COLUMN OPPRETTET_AV SET NOT NULL;
+ALTER TABLE PROSESS_TASK_FEILHAND
+  ALTER COLUMN NAVN SET NOT NULL;
+ALTER TABLE PROSESS_TASK_FEILHAND
+  ALTER COLUMN KODE SET NOT NULL;
+
+ALTER TABLE PROSESS_TASK_TYPE
+  ADD CONSTRAINT PK_PROSESS_TASK_TYPE PRIMARY KEY (KODE);
+ALTER TABLE PROSESS_TASK_TYPE
+  ALTER COLUMN OPPRETTET_TID SET NOT NULL;
+ALTER TABLE PROSESS_TASK_TYPE
+  ALTER COLUMN OPPRETTET_AV SET NOT NULL;
+ALTER TABLE PROSESS_TASK_TYPE
+  ALTER COLUMN FEIL_SEK_MELLOM_FORSOEK SET NOT NULL;
+ALTER TABLE PROSESS_TASK_TYPE
+  ALTER COLUMN FEIL_MAKS_FORSOEK SET NOT NULL;
+ALTER TABLE PROSESS_TASK_TYPE
+  ALTER COLUMN KODE SET NOT NULL;
+
+
+ALTER TABLE PROSESS_TASK
+  ADD CONSTRAINT FK_PROSESS_TASK_1 FOREIGN KEY (TASK_TYPE)
+    REFERENCES PROSESS_TASK_TYPE (KODE);
+
+ALTER TABLE PROSESS_TASK_TYPE
+  ADD CONSTRAINT FK_PROSESS_TASK_TYPE_1 FOREIGN KEY (FEILHANDTERING_ALGORITME)
+    REFERENCES PROSESS_TASK_FEILHAND (KODE);
