+++ web/webapp/pom.xml
@@ -54,9 +54,10 @@
         </dependency>
 
         <!-- NAV -->
-
-        <!-- Server Provided Libraries. Transitive avhengigheter som ikke burde
-              vært her da de trekkes inn av server. -->
+        <dependency>
+            <groupId>no.nav.foreldrepenger.felles</groupId>
+            <artifactId>felles-prosesstask</artifactId>
+        </dependency>
         <dependency>
             <groupId>no.nav.foreldrepenger.felles.sikkerhet</groupId>
             <artifactId>felles-sikkerhet</artifactId>
